bumped to Titan 0.4.0-SNAPSHOT and TinkerPop 2.4.0. Fixed logging issue that resulted from Gremlin upgrade.

+++ src/test/java/com/thinkaurelius/faunus/formats/VertexQueryFilterTest.java
@@ -18,7 +18,7 @@ public class VertexQueryFilterTest extends BaseTest {
         Configuration config = new Configuration();
         VertexQueryFilter query = VertexQueryFilter.create(config);
         assertFalse(query.doesFilter());
-        assertEquals(query.limit, Long.MAX_VALUE);
+        assertEquals(query.limit, Integer.MAX_VALUE);
         assertEquals(query.hasContainers.size(), 0);
         assertEquals(query.direction, Direction.BOTH);
         assertEquals(query.labels.length, 0);
